Adds private constructors to static interfaces.
All of the NaCl and Libsodium interfaces are static so there is no
reason to ever instantiate any of the classes that the methods are
organized under. Added private constructors to all of them to make that
clear.

Inlcudes a global reformat to pickup formating added by the maven
javadoc command.